Added Projectile Class and some fixes
- Projectile.cs
- BoomStick.cs inherits from Projectile.cs
- Some temporary fixes to Room.cs (collision detection and dynamic
addition/removal of objects)
- Added a few sample textures for particles.
- Players shoot with left and right ctrl buttons now, but collision
detection for the projectiles is still buggy and needs proper fixage

+++ CrashNet/CrashNet/ParticleSystem/ParticleEmitter.cs
@@ -338,10 +338,10 @@ namespace CrashNet.ParticleSystem
                 (int)(GlobalParticleLevels * Convert.ToInt32(emitterPars.SelectSingleNode("emitRate").
                 Attributes.GetNamedItem("x").Value));
             MeanEmitDelay =
-                Convert.ToInt32(emitterPars.SelectSingleNode("meanEmitDelay").
+                Convert.ToDouble(emitterPars.SelectSingleNode("meanEmitDelay").
                 Attributes.GetNamedItem("x").Value);
             EmitLifetime =
-                Convert.ToInt32(emitterPars.SelectSingleNode("emitterLifetime").
+                Convert.ToDouble(emitterPars.SelectSingleNode("emitterLifetime").
                 Attributes.GetNamedItem("x").Value);
             PermanentParticles =
                 Convert.ToBoolean(emitterPars.SelectSingleNode("permanentParticles").
